% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{avg_dists}
\alias{avg_dists}
\title{Average Pairwise Distance within Subgraphs}
\usage{
avg_dists(G)
}
\arguments{
\item{G}{A graph whose subgraphs we want to find.}
}
\description{
Average Pairwise Distance within Subgraphs
}
