% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{background_prop}
\alias{background_prop}
\title{Proportion of image that's Background}
\usage{
background_prop(x, ...)
}
\description{
This is an example of a function that can be called in \code{loop_stats}.
}
