% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{cell_type}
\alias{cell_type}
\title{Helper to get Cell Type}
\usage{
cell_type(exper)
}
\arguments{
\item{exper}{A SummarizedExperiment representation of the MIBI data.}
}
\description{
Combines the tumor and immune group variables into a generic cell type.
}
