% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{data_list}
\alias{data_list}
\title{Wrapper to get SCE data}
\usage{
data_list(pattern)
}
\arguments{
\item{pattern}{A partial matching string to use to extract variables from the
global environment.}
}
\description{
Wrapper to get SCE data
}
