% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{download_data}
\alias{download_data}
\title{Download Data for Experiments}
\usage{
download_data(directory)
}
\arguments{
\item{directory}{Directory to which all the data is downloaded.}
}
\description{
Attempts to download data to directory if it's not already found.
}
