% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_graph}
\alias{extract_graph}
\title{Extract a KNN Graph from Polygons}
\usage{
extract_graph(geometries, K = 5)
}
\arguments{
\item{geometries}{The data.frame containing the polygons and their labels.}

\item{K}{The number of nearest neighbors to which each polygon will be
linked.}
}
\description{
Sometimes we want to put neighboring polygons into one big graph.
}
