% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{fit_wrapper}
\alias{fit_wrapper}
\title{Helper to Fit Models}
\usage{
fit_wrapper(x, y)
}
\arguments{
\item{x}{The covariates to use for training.}

\item{y}{The response to fit to.}
}
\description{
Helper to Fit Models
}
