% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{generate_model}
\alias{generate_model}
\title{importFrom keras keras_model_sequential layer_dense layer_activation}
\usage{
generate_model(n_ft)
}
\description{
importFrom keras keras_model_sequential layer_dense layer_activation
}
