% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{graph_stats_cell}
\alias{graph_stats_cell}
\title{Apply fun to Graph Neighborhoods}
\usage{
graph_stats_cell(cell_id, G, polys, fun, ...)
}
\arguments{
\item{cell_id}{The ID of the cell to extract a local neighborhood around.}

\item{G}{The graph object giving the connections between cell_ids.}

\item{polys}{A spatial data.frame with a column (geometry) giving the spatial
geometry of each cell.}

\item{fun}{A function that can be applied to a data.frame whose rows are
pixels and whose columns give features of those pixels (e.g., immune
group).}
}
\value{
result A tibble mapping the cell to statistics calculated by fun.
}
\description{
Apply fun to Graph Neighborhoods
}
