% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_views.R
\name{linked_views}
\alias{linked_views}
\title{Interactive Linked_Viewss}
\usage{
linked_views(polys, dimred, width = 900, height = 450, elementId = NULL)
}
\arguments{
\item{polys}{A geojson object containing the cell coordinates.}

\item{dimred}{A dataframe with columns V1 and V2 containing the U-Map layout
of the cells.}

\item{width}{The width of the output visualization.}

\item{height}{The width of the output visualization.}

\item{elementId}{The ID of the new htmlwidgets element within which to put
the vis.}
}
\description{
Interactive Linked_Viewss
}
