% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{load_mibi}
\alias{load_mibi}
\title{Load MIBI-ToF into List}
\usage{
load_mibi(data_dir, n_paths = NULL)
}
\arguments{
\item{data_dir}{The directory containing the MIBI single cell binary as well
as a directory containing all rasters.}

\item{n_paths}{If you want to subsample to a small number of images, set the
limit to n_paths.}
}
\description{
Helper function to load MIBI-TOF data.
}
