% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{loop_stats}
\alias{loop_stats}
\title{Wrapper for Local Statistics}
\usage{
loop_stats(cell_ids, type = "raster", ...)
}
\arguments{
\item{cell_ids}{A vector of cell IDs on which to apply a function to}

\item{type}{Either "raster" or "graph". Specifies the types of neighborhoods
(image or graph) on which to compute statistics.}

\item{...}{Additional arguments to the cell_fun.}
}
\description{
Loop functions that can be applied on a cell by cell basis.
}
