% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{raster_stats_cell}
\alias{raster_stats_cell}
\title{Apply fun to Local Neighborhoods}
\usage{
raster_stats_cell(
  cell_id,
  im,
  polys,
  fun,
  buffer_radius = 90,
  plot_masks = TRUE
)
}
\arguments{
\item{cell_id}{The ID of the cell to extract a local neighborhood around.}

\item{im}{The raster object giving the pixel-level information about the
sample.}

\item{polys}{A spatial data.frame with a column (geometry) giving the spatial
geometry of each cell.}

\item{fun}{A function that can be applied to a data.frame whose rows are
pixels and whose columns give features of those pixels (e.g., immune
group).}

\item{buffer_radius}{The size of the window around cell_id, to use to subset
the raster on which to apply fun.}

\item{plot_masks}{If you want to see what the subsets of cells looks like,
you can use this.}
}
\value{
result A tibble mapping the cell to statistics calculated by fun.
}
\description{
Apply fun to Local Neighborhoods
}
