% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{sample_proportions}
\alias{sample_proportions}
\title{Division of Samples into Clusters}
\usage{
sample_proportions(SampleID, cluster)
}
\arguments{
\item{SampleID}{The sample IDs associated with each cell.}

\item{cluster}{The cluster membership of each cell.}
}
\description{
Get each samples' distribution across clusters.
}
