% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{spatial_subsample}
\alias{spatial_subsample}
\title{Subsample Rasters}
\usage{
spatial_subsample(tiff_paths, exper, qsize = 500)
}
\arguments{
\item{tiff_paths}{The paths containing the full raster images.}

\item{exper}{A summarized experiment object, containing measurements for each
cell.}

\item{qsize}{How large should the rasters be?}
}
\description{
This lets you subsample a large raster image and the associated experiment
object, for quicker experimentation.
}
