% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{subgraphs}
\alias{subgraphs}
\title{All Subgraphs of a Graph}
\usage{
subgraphs(G, order = 3)
}
\arguments{
\item{G}{A graph whose subgraphs we want to find.}

\item{order}{The size of the subgraphs.}
}
\description{
All Subgraphs of a Graph
}
