% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{type_props}
\alias{type_props}
\title{Proportion of Cell Types}
\usage{
type_props(x, ...)
}
\arguments{
\item{x}{A data frame containing cell type information (in a column called
$cell_type)}

\item{...}{Additional unused arguments.}
}
\description{
Function to extract proportion of cell types within each subgraph.
}
